adds put and delete

+++ src/main/java/com/bbs/CustomerNotFoundException.java
@@ -2,7 +2,7 @@ package com.bbs;
 
 public class CustomerNotFoundException extends RuntimeException {
 
-	CustomerNotFoundException(int id) {
+	CustomerNotFoundException(long id) {
 		    super("Could not find customer " + id);
 		  }
