+++ src/main/java/com/bbs/CustomerRepository.java
@@ -2,7 +2,7 @@ package com.bbs;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface CustomerRepository extends CrudRepository<Customer, Integer> {
+public interface CustomerRepository extends CrudRepository<Customer, Long> {
